claude-agent/requirements.txt
# Claude Agent SDK and dependencies
claude-agent-sdk>=0.1.18,<0.2.0
python-dotenv>=1.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
